[BUGFIX] Make global scope available when running in Node.js land.
This PR contains a bugfix that resolves an `undefined` passed global variable
when running in a Node.js execution environment.

The problem was that the defined variable `global` overwrites the real "global" object
which causes an unmatched condition in the statement:

...

else if (typeof global !== "undefined") {
    global = global;
...

`global` was in this case always `undefined`. This problem only occurs in the "unminified" artefact.
Anyway, this is a critical problem though. The tests (e.g. `assembler`) utilizes the unminified files.